# frontend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0
  description: ユーザー管理APIのドキュメント
  contact:
    name: APIサポート
    url: http://www.example.com/support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
...
  /admin/auth/login:
    post:
      tags:
        - Admin
      summary: ログイン
      description: ログインします。
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: ログイン成功
        '400':
          description: ログイン失敗
...